Hook shutdown event / Override network PORT
Tested on Win32. 
Not sure if this method works on Linux, so added code in ELSE clause of #def __linux__.

Hooking the shutdown event allows the code to gracefully shutdown a connection to a device or properly dispose of links to managed code.

This commit also defaults the network PORT to 8080 (POCO default of 80 is typically blocked by Windows). An optional command line parameter can be used to override the PORT number.